requirements.txt
Pillow~=9.2.0
qrcode~=7.3.1
requests~=2.28.1